Combine APIs into api.py
